Block model review comments (part 2)
