+++ flamedisx/lxe_blocks/detection.py
@@ -96,8 +96,7 @@ class DetectPhotons(DetectPhotonsOrElectrons):
     dimensions = ('photons_produced', 'photons_detected')
 
     special_model_functions = ('photon_acceptance', 'penning_quenching_eff')
-    model_functions = tuple(['photon_detection_eff']
-                            + list(special_model_functions))
+    model_functions = ('photon_detection_eff',) + special_model_functions
 
     photon_detection_eff = 0.1
 
@@ -126,8 +125,7 @@ class DetectElectrons(DetectPhotonsOrElectrons):
     dimensions = ('electrons_produced', 'electrons_detected')
 
     special_model_functions = ('electron_acceptance',)
-    model_functions = tuple(['electron_detection_eff'] +
-                            list(special_model_functions))
+    model_functions = ('electron_detection_eff',) + special_model_functions
 
     @staticmethod
     def electron_detection_eff(drift_time, *,
